fix some warnings
git-svn-id: https://svn.code.sf.net/p/pinmame/code/trunk@4236 6f51dfdd-75df-4273-8f3b-818b1e6aa8bf

+++ src/sound/disc_inp.c
@@ -50,7 +50,7 @@ READ_HANDLER(discrete_sound_r)
 	/* Update the node input value if allowed */
 	if(dss_input_map[offset])
 	{
-		data=dss_input_map[offset]->input[0];
+		data=(int)dss_input_map[offset]->input[0];
 	}
     return data;
 }
